requirements.txt
python-dotenv==1.0.1
pyttsx3==2.90
pyautogui==0.9.54
keyboard==0.13.5
mouse==0.7.1
Pillow==10.4.0
requests==2.32.3
websocket-client==1.8.0
SpeechRecognition==3.10.4
pyaudio==0.2.14
vosk==0.3.45
playwright==1.48.0
uiautomation==2.0.20
PyGetWindow==0.0.9
pycaw==20240210
pytesseract==0.3.13
numpy<2
opencv-python-headless==4.10.0.84
audioop-lts==0.2.2
spacy>=3.7.0
pyperclip
openai>=1.0.0
google-generativeai>=0.3.0
spotipy>=2.23.0
psutil>=5.9.0
